refactor(stim): consolidate ResolveSet{Observable,Detector} into ResolveSetAnnotate
The two resolvers shared their entire GetRecord-chain construction — frame
lookup, MeasureIdTuple validation, num_measures_at_stmt check, and the
per-element py.Constant + GetRecord emission loop. Only the wrapping stmt
differed (ObservableInclude vs Detector) plus their two unique bits
(observable_count counter vs extract_coord_ssas call).

Consolidates both into a single ResolveSetAnnotate rule with the shared
logic extracted into a _build_get_record_chain helper. Follows the
existing precedent of ResolveGetRecIdx, which already handles both
GetRecIdxFromMeasurement and GetRecIdxFromPredicate in one rule.

Net: ~30 lines removed, one fewer Walk() in the step-10 pipeline chain,
no behavioral change.
