+++ src/bloqade/stim/rewrite/__init__.py
@@ -8,7 +8,6 @@ from .scf_for_to_repeat import ScfForToRepeat as ScfForToRepeat
 from .ifs_to_stim_partial import IfToStimPartial as IfToStimPartial
 from .py_constant_to_stim import PyConstantToStim as PyConstantToStim
 from .resolve_get_rec_idx import ResolveGetRecIdx as ResolveGetRecIdx
-from .resolve_set_detector import ResolveSetDetector as ResolveSetDetector
+from .resolve_set_annotate import ResolveSetAnnotate as ResolveSetAnnotate
 from .set_detector_partial import SetDetectorPartial as SetDetectorPartial
-from .resolve_set_observable import ResolveSetObservable as ResolveSetObservable
 from .set_observable_partial import SetObservablePartial as SetObservablePartial
